// app/types.ts
export interface IPontoRota {
  id: string;
  nome: string;
  ordem: number;
  horario_chegada?: string;
  horario_partida?: string;
  tipo: 'ORIGEM' | 'PARADA_INTERMEDIARIA' | 'DESTINO';
  permite_embarque: boolean;
  permite_desembarque: boolean;
  observacoes?: string;
  state_id?: number;
  city_id?: number;
  neighborhood_id?: number;
  distancia_do_anterior_km?: number;
  duracao_deslocamento_minutos?: number;
  duracao_parada_minutos?: number;
  tempo_acumulado_minutos?: number;
}